specialize `copyto!` for performance
The fallback generic implementation of `copyto!` uses `setindex!` in an
elementwise manner. Instead base the implementation on
`unsafe_copyto!`. This improves the performance of `copyto!`, thus it
also improves the performance of `copy!`, `copy` and of broadcasted
operations.

This change only applies when copying between two `FixedSizeArray`s,
or between a `Memory` and a `FixedSizeArray`.

Fixes #19